Fix: 타입정리
